// src/api/interface/menuIterface.ts
export interface TypeObject {
  id: number
  name: string
  description: string
}

export interface MenuDataType {
  id: number
  type: TypeObject
  title: string
  availbility?: string // Typo in property name, should be corrected if intentional
  availability: string
  sku: string
  brand: string
  rating: number
  price: number
  condition: string
  description: string
  category: string
  link: string
  image_link: string
  restaurant?: number
  business?: number
  name?: string
  status?: string
  availability_circle_radius: number
  imageUrl?: string
  size?: string
  topping?: string
  // quantity: string
  quantity: number
}